Add functionality to upload an image
Allow users to upload an image from their file system. This image will
be displayed on the page. When an image is uploaded, an animated
placeholder is also displayed. This represents where a processed image
will appear.

Also log a statement to the console when an image is clicked about the
color of the clicked pixel.