Add prod flag to the Angular build that Heroku runs
When our application is deployed to Heroku, Heroku looks at the
root-level package.json for a "heroku-postbuild" script. Presently, this
script is configured to run "npm run build" within the web/ directory.
This command runs a build based on the "build" script within
web/package.json.

In commit d613954, the --prod flag was inadvertently removed from
web/package.json. In addition, a second "heroku-postbuild" script was
added, but it was not needed. This change set reinstates
web/package.json to have the --prod flag on the build script and also
removes this unneeded heroku-postbuild script.