Upload images from the frontend in base64 encoding
Upload images from the ImageService with a form containing the base64
representation of an image, as well as the image's name. On the backend,
decode the image before processing and returning it as an encoded image.
When the processed, encoded image is returned, decode it once again on
the frontend.