Clean up ImageService test suite
Stop testing the uploadImage() method in the fakeAsync zone. In commit
bbd7225, we changed the responseType of the /upload API from blob to
JSON. We previously tested this method as fakeAsync because it was
required for this response type. As the response type has changed, the
test can be simplified to remove this code.