Define a download API
Expose a download endpoint that, when called, returns up to four image
files (original, original with outline, segmented, segmented with
outline), depending on the request body. The image(s) to be returned
match the serialID in the request body. Return the response as a ZIP
file and then perform cleanup to remove the images.