Check for allowed file extensions based on the original filename
Previously, we were checking if an uploaded file's extension was allowed
based on a generated filename. This generated filename was a combination
of SerialId + '.png'. In other words, we were always checking if a PNG
file was allowed, regardless of whether the uploaded file was actually
in PNG format or not.